Add and Delete function with Image in Accommodation Page

+++ app/Http/Controllers/RatesController.php
@@ -2,11 +2,9 @@
 
 namespace App\Http\Controllers;
 
-use Yajra\DataTables\DataTables;
 use App\Models\Fare;
 use Illuminate\Http\Request;
-
-
+use Yajra\DataTables\Facades\DataTables;
 
 class RatesController extends Controller
 {
@@ -21,19 +19,20 @@ class RatesController extends Controller
             $data = Fare::latest()->get();
             return DataTables::of($data)->addColumn('action', function ($fare) {
                 $editUrl = route('rates.edit', $fare->id);
+                $viewUrl = route('rates.show', $fare->id);
                 $deleteUrl = route('rates.destroy', $fare->id);
 
-                $editButton = '<a href="' . $editUrl . '" class="btn btn-primary btn-sm edit" data-toggle="modal" data-target="#editModal" data-id="' . $fare->id . '">Edit</a>';
-                $deleteButton = '<button type="button" class="btn btn-danger btn-sm delete" data-url="' . $deleteUrl . '">Delete</button>';
+                $editButton = '<a href="' . $editUrl . '" class="btn btn-primary btn-sm edit" data-toggle="modal" data-target="#editModal" data-id="' . $fare->id . '"><i class="fa fa-edit"></i>Edit</a>';
+                $viewButton = '<a href="' . $viewUrl . '" class="btn btn-success btn-sm view-btn" data-toggle="modal" data-target="#viewModals" data-id="' . $fare->id . '"><i class="fa fa-eye"></i>View</a>';
+                $deleteButton = '<button type="button" class="btn btn-danger btn-sm delete" data-url="' . $deleteUrl . '" data-id="' . $fare->id . '"><i class="fa fa-trash"></i>Delete</button>';
 
-                return $editButton . ' ' . $deleteButton;
+                return $editButton . ' ' . $viewButton . ' ' . $deleteButton;
             })
                 ->rawColumns(['action'])
                 ->make(true);
         }
         return view('admin.settings.rates.index');
     }
-
     /**
      * Show the form for creating a new resource.
      *
@@ -41,6 +40,7 @@ class RatesController extends Controller
      */
     public function create()
     {
+        return view('dashboard.settings.rates.create');
         return view('dashboard.settings.rates.create');
     }
 
@@ -67,6 +67,21 @@ class RatesController extends Controller
         } else {
             return response()->json(['error' => 'Failed to add rates.']);
         }
+        $request->validate([
+            'fare_name' => 'required|string|max:255',
+            'price' => 'required|numeric',
+        ]);
+
+        $rates = Fare::create([
+            'fare_name' => $request->fare_name,
+            'price' => $request->price,
+        ]);
+
+        if ($rates) {
+            return response()->json(['success' => 'Rates added successfully.']);
+        } else {
+            return response()->json(['error' => 'Failed to add rates.']);
+        }
     }
 
     /**
@@ -77,7 +92,11 @@ class RatesController extends Controller
      */
     public function show($id)
     {
-        //
+        $fare = Fare::find($id);
+
+        return response()->json([
+            'fare' => $fare,
+        ])->header('Content-Type', 'application/json');
     }
 
     /**
@@ -88,6 +107,8 @@ class RatesController extends Controller
      */
     public function edit($id)
     {
+        $fare = Fare::find($id);
+        return response()->json($fare);
         $fare = Fare::find($id);
         return response()->json($fare);
     }
@@ -116,6 +137,21 @@ class RatesController extends Controller
         } else {
             return response()->json(['error' => 'Rates not found.']);
         }
+        $fare = Fare::find($id);
+
+        if ($fare) {
+            $request->validate([
+                'fare_names' => 'required|string|max:255',
+                'prices' => 'required|numeric',
+            ]);
+
+            $fare->fare_name = $request->fare_names;
+            $fare->price = $request->prices;
+            $fare->save();
+            return response()->json(['success' => 'Rates updated successfully.']);
+        } else {
+            return response()->json(['error' => 'Rates not found.']);
+        }
     }
 
     /**
@@ -124,10 +160,32 @@ class RatesController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
+    // public function destroy(Request $request, $id)
+    // {
+    //     if ($request->ajax()) {
+    //         $fare = Fare::find($id);
+    //         if ($fare) {
+    //             $fare->delete();
+    //             return response()->json(['success' => true, 'message' => 'Rates deleted successfully']);
+    //         } else {
+    //             return response()->json(['success' => false, 'message' => 'Rates not found']);
+    //         }
+    //     } else {
+    //         return redirect()->route('rates.index')->with('error', 'Invalid request');
+    //     }
+    // }
     public function destroy(Request $request, $id)
     {
-        $fare = Fare::findOrFail($id);
-        $fare->delete();
-        return response()->json(['success' => 'Rates deleted successfully.']);
+        if ($request->ajax()) {
+            $fare = Fare::findOrFail($id);
+            if ($fare) {
+                $fare->delete();
+                return response()->json(['success' => true, 'message' => 'Rates deleted successfully']);
+            } else {
+                return response()->json(['success' => false, 'message' => 'Rates not found']);
+            }
+        } else {
+            return redirect()->route('accomodations.index')->with('error', 'Invalid request');
+        }
     }
 }
